requirements.txt
pillow>=9.0.0
pytesseract>=0.3.10
pyautogui>=0.9.54
opencv-python>=4.5.0
easyocr>=1.6.0
numpy>=1.21.0
pyperclip>=1.8.0
scikit-image>=0.19.0
